+++ src/Storage/Storage.Management/StorageAccount/NewAzureStorageAccount.cs
@@ -249,6 +249,23 @@ namespace Microsoft.Azure.Commands.Management.Storage
 
         [Parameter(
             Mandatory = false,
+            HelpMessage = "Enables extended group support with local users feature, if set to true.")]
+        [ValidateNotNullOrEmpty]
+        public bool EnableExtendedGroup
+        {
+            get
+            {
+                return enableExtendedGroup != null ? enableExtendedGroup.Value : false;
+            }
+            set
+            {
+                enableExtendedGroup = value;
+            }
+        }
+        private bool? enableExtendedGroup = null;
+
+        [Parameter(
+            Mandatory = false,
             HelpMessage = "Enable HierarchicalNamespace for the Storage account.")]
         [ValidateNotNullOrEmpty]
         public bool EnableHierarchicalNamespace
@@ -914,6 +931,10 @@ namespace Microsoft.Azure.Commands.Management.Storage
             if (this.enableLocalUser != null)
             {
                 createParameters.IsLocalUserEnabled = this.enableLocalUser;
+            }
+            if (this.enableExtendedGroup != null)
+            {
+                createParameters.EnableExtendedGroups = this.enableExtendedGroup;
             }
             if (this.AllowedCopyScope != null)
             {